OutfitMatch/requirements.txt
beautifulsoup4==4.13.4
config==0.5.1
Flask==3.1.1
flask_cors==6.0.0
neo4j==5.28.1
numpy<2
openai==1.83.0
pandas==2.2.3
Pillow==11.2.1
psycopg2==2.9.10
python-dotenv==1.1.0
Requests==2.32.3
scikit_learn==1.6.1
selenium==4.33.0
torch>=2.0.0
torchvision>=0.15.0
transformers==4.52.4
webdriver_manager==4.0.2